backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
django-filter==23.1
djoser==2.1.0
django-colorfield==0.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
gunicorn==20.1.0